requirements.txt
astropy==6.1.7
Cartopy==0.24.1
geopandas==1.0.1
glfw==2.9.0
imgui==2.0.0
matplotlib==3.10.3
numpy==2.2.6
Pillow==11.2.1
PyOpenGL==3.1.9
PyYAML==6.0.2
Requests==2.32.3
Shapely==2.1.1
tqdm==4.66.4
rustworkx
pandas
seaborn
scipy
plotly

--extra-index-url https://download.pytorch.org/whl/cu128
torch
torchvision
torchaudio
tensorboard